# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sanjeevani Backend",
    "description": "Sanjeevani supply-chain disruption system backend.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/calls/start": {
      "post": {
        "tags": [
          "calls"
        ],
        "summary": "Start Call",
        "operationId": "start_call_api_v1_calls_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calls/{call_id}": {
      "get": {
        "tags": [
          "calls"
        ],
        "summary": "Get Call",
        "operationId": "get_call_api_v1_calls__call_id__get",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calls/{call_id}/replay": {
      "post": {
        "tags": [
          "calls"
        ],
        "summary": "Replay Call",
        "operationId": "replay_call_api_v1_calls__call_id__replay_post",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/bolna": {
      "post": {
        "tags": [
          "calls"
        ],
        "summary": "Receive Bolna Webhook",
        "operationId": "receive_bolna_webhook_api_v1_webhooks_bolna_post",
        "parameters": [
          {
            "name": "x-voice-adapter-secret",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Voice-Adapter-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CallSessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "disruption_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disruption Id"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "briefing_snapshot": {
            "type": "object",
            "title": "Briefing Snapshot"
          },
          "guardrails": {
            "type": "object",
            "title": "Guardrails"
          },
          "transcript": {
            "items": {},
            "type": "array",
            "title": "Transcript"
          },
          "extracted": {
            "type": "object",
            "title": "Extracted"
          },
          "validation": {
            "type": "object",
            "title": "Validation"
          },
          "correlation_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Method"
          },
          "outcome_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "disruption_id",
          "vendor_id",
          "status",
          "source",
          "started_at",
          "briefing_snapshot",
          "guardrails",
          "transcript",
          "extracted",
          "validation"
        ],
        "title": "CallSessionResponse"
      },
      "CallStartRequest": {
        "properties": {
          "disruption_id": {
            "type": "string",
            "title": "Disruption Id"
          },
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "mode": {
            "type": "string",
            "pattern": "^(LIVE|REPLAY)$",
            "title": "Mode",
            "default": "LIVE"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "disruption_id",
          "vendor_id"
        ],
        "title": "CallStartRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}